Move to make-based build process
Fixed the tentative makefile and added a linker script so that building from source should now be less of a pain for people who clone this repository. Also slightly modified the source to account for this (primarily the fact that handling the format of the .COM executable is now the responsibility of the linker script, which should work better if I ever need another source file).